CRUD operations and related tests are added

+++ src/main/java/mb/oauth2authorizationserver/data/repository/UserRepository.java
@@ -3,7 +3,7 @@ package mb.oauth2authorizationserver.data.repository;
 import mb.oauth2authorizationserver.data.entity.SecurityUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<SecurityUser, Integer> {
+public interface UserRepository extends JpaRepository<SecurityUser, Long> {
 
     SecurityUser findByUsername(String username);
 }
